requirements.txt
streamlit
requests
groq
python-dotenv